src/mini_solver_backend/requirements.txt
python-dotenv

pydantic
aio-pika
kink

firebase-admin
google-cloud-firestore
google-cloud-storage
google-auth
google-auth-oauthlib